refactor history view into smaller views
