+++ CosmoPic/Views/FavoritesView.swift
@@ -43,7 +43,3 @@ struct FavoritesView: View {
     }
   }
 }
-
-#Preview {
-  FavoritesView()
-}
